+++ CHANGELOG.md
@@ -22,6 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added a `line` marker for drawing arbitrary lines on a board.
 * In the `stacking-tiles` renderer, the click handler now returns the number of the tile being clicked, with 1 being the top piece.
 * Added `hexFill` property that will fill hexes in a hex grid with a given colour.
+* Added `width` property to the `fence` marker so the game can adjust the weight of the fence.
+* Added a `clickable-edges` option that only affects rect-of-hex maps. It renders the edges using the base stroke width, colour, etc., but it also attaches a click handler, returning the row, column, and face the edge represents. The renderer does not double up on edges (e.g., a1NE and b1SW are the same edge, and only one will be rendered).
+* Added glyphs for Streetcar Suburb.
 
 ### Fixed